fix: adjust tool registration for IntelligenceHandler
