+++ packages/qdrant-loader-core/src/qdrant_loader_core/graph/extractor/confluence.py
@@ -40,6 +40,30 @@ class ConfluenceEntityExtractor(BaseEntityExtractor):
     ) -> str | None:
         return doc.metadata.get("space_key")
 
+    def _build_document_node(
+        self,
+        doc: Document,
+        project: str | None,
+    ) -> GraphNode:
+        """
+        Build a document node enriched with filesystem metadata.
+        """
+
+        file_name = doc.metadata.get("file_name")
+
+        return GraphNode(
+            id=doc.id,
+            label=CoreNodeLabel.DOCUMENT.value,
+            project=project,
+            properties={
+                "title": file_name,
+                "url": doc.url,
+                "created_at": doc.metadata.get("created_at"),
+                "updated_at": doc.metadata.get("updated_at"),
+                "source_type": self.source_type,
+            },
+        )
+
     # ------------------------------------------------------------------
     # People
     # ------------------------------------------------------------------
