+++ packages/qdrant-loader-mcp-server/src/qdrant_loader_mcp_server/mcp/intelligence_handler.py
@@ -12,6 +12,7 @@ from qdrant_loader_core.graph import get_graph_store
 from ..search.engine import SearchEngine
 from ..utils import LoggingConfig
 from .formatters import MCPFormatters
+from .graph_handler import handle_find_ticket_dependencies
 from .handlers.intelligence import (
     get_or_create_document_id as _get_or_create_document_id_fn,
 )
@@ -983,98 +984,5 @@ class IntelligenceHandler:
             raise ValueError("depth must be between 1 and 10")
         return depth
 
-    async def handle_find_ticket_dependencies(
-        self, request_id: str | int | None, params: dict[str, Any]
-    ):
-        """
-        Traverse Jira blocking dependencies
-        """
 
-        if "ticket_key" not in params:
-            logger.error("Missing required parameter: ticket_key")
-            return self.protocol.create_response(
-                request_id,
-                error={
-                    "code": -32602,
-                    "message": "Invalid params",
-                    "data": "Missing required parameter: ticket_key",
-                },
-            )
-
-        if "depth" not in params:
-            logger.error("Missing required parameter: depth")
-            return self.protocol.create_response(
-                request_id, error={"code": -32602, "message": "Invalid params"}
-            )
-
-        depth = params.get("depth")
-        ticket_key = params.get("ticket_key")
-
-        if not isinstance(depth, int):
-            logger.error("Invalid depth parameter type")
-            return self.protocol.create_response(
-                request_id,
-                error={
-                    "code": -32602,
-                    "message": "Invalid params",
-                    "data": "depth must be an integer",
-                },
-            )
-
-        query = f"""
-        MATCH path =
-            (start:Document {{id: $ticket_key}})
-            -[:LINKS_TO*1..{depth}]->
-            (target)
-        RETURN
-            nodes(path),
-            relationships(path)
-        """
-
-        query_params = {
-            "ticket_key": ticket_key,
-            "depth": depth,
-        }
-        try:
-            depth = self._validate_depth(depth)
-
-            result = await self._run_graph_query(query, query_params)
-
-            formatted = self.formatters.format_graph(result)
-
-            return self.protocol.create_response(
-                request_id,
-                result={
-                    "content": [
-                        {
-                            "type": "text",
-                            "text": (
-                                f"Found {len(formatted.get('nodes', []))} "
-                                f"nodes and {len(formatted.get('edges', []))} edges"
-                            ),
-                        }
-                    ],
-                    "structuredContent": formatted,
-                    "isError": False,
-                },
-            )
-
-        except ValueError as e:
-            return self.protocol.create_response(
-                request_id,
-                error={
-                    "code": -32602,
-                    "message": "Invalid params",
-                    "data": str(e),
-                },
-            )
-
-        except Exception:
-            logger.exception("Error querying graph")
-            return self.protocol.create_response(
-                request_id,
-                error={
-                    "code": -32603,
-                    "message": "Internal server error",
-                },
-            )
+IntelligenceHandler.handle_find_ticket_dependencies = handle_find_ticket_dependencies
